# docs/swagger.yaml
definitions:
  problemdetail.ProblemDetail:
    properties:
      context: {}
      detail:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/test-plans:
    get:
      consumes:
      - application/json
      description: Search all Test Plans
      operationId: SearchTestPlans
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Search all Test Plans
      tags:
      - test-plans
    post:
      consumes:
      - application/json
      description: Create a new Test Plan
      operationId: CreateTestPlan
      parameters:
      - description: Create Test plan data
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Create a new Test Plan
      tags:
      - test-plans
  /api/v1/test-plans/{testPlanID}:
    delete:
      consumes:
      - application/json
      description: Delete a Test Plan
      operationId: DeleteTestPlan
      parameters:
      - description: Test Plan ID
        in: path
        name: testPlanID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Delete a Test Plan
      tags:
      - test-plans
    get:
      consumes:
      - application/json
      description: Get one Test Plan
      operationId: GetOneTestPlan
      parameters:
      - description: Test Plan ID
        in: path
        name: testPlanID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Get one Test Plan
      tags:
      - test-plans
    post:
      consumes:
      - application/json
      description: Update a Test Plan
      operationId: UpdateTestPlan
      parameters:
      - description: Test Plan ID
        in: path
        name: testPlanID
        required: true
        type: string
      - description: Test Plan data
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Update a Test Plan
      tags:
      - test-plans
  /api/v1/test-runs:
    get:
      consumes:
      - application/json
      description: List Test Runs
      operationId: ListTestRuns
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: List Test Runs
      tags:
      - test-runs
    post:
      consumes:
      - application/json
      description: Create a new Test Run
      operationId: CreateTestRun
      parameters:
      - description: Test Run data
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Create a new Test Run
      tags:
      - test-runs
  /api/v1/test-runs/{testRunID}:
    delete:
      consumes:
      - application/json
      description: Delete a Test Run
      operationId: DeleteTestRun
      parameters:
      - description: Test Run ID
        in: path
        name: testRunID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Delete a Test Run
      tags:
      - test-runs
    get:
      consumes:
      - application/json
      description: Get one Test Run
      operationId: GetOneTestRun
      parameters:
      - description: Test Run ID
        in: path
        name: testRunID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Get one Test Run
      tags:
      - test-runs
    post:
      consumes:
      - application/json
      description: Update a Test Run
      operationId: UpdateTestRun
      parameters:
      - description: Test Run ID
        in: path
        name: testRunID
        required: true
        type: string
      - description: Test Run update data
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Update a Test Run
      tags:
      - test-runs
  /api/v1/test-runs/query:
    get:
      consumes:
      - application/json
      description: Search for Test Runs
      operationId: SearchTestRuns
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Search for Test Runs
      tags:
      - test-runs
...
swagger: "2.0"